Populate link address cache at dispatch
This allows the stack to learn remote link addresses on incoming
packets, reducing the need to ARP to send responses.

Modify the cache to use a read-write lock to reduce contention, since
calls to `add` will be much more frequent now.

Reduce the number of round trips to the system clock while I'm here,
since that may also prove to be performance-sensitive.

Fixes #739.

PiperOrigin-RevId: 264702449